Ajuste na classe listagemVIEW e ProdutosDAO para que os produtos sejam listados na tabela

+++ src/ProdutosDAO.java
@@ -65,7 +65,7 @@ public class ProdutosDAO {
                 ProdutosDTO produto = new ProdutosDTO();
                 produto.setId(resultset.getInt("id"));
                 produto.setNome(resultset.getString("nome"));
-                produto.setValor(resultset.getInt("valor"));
+                produto.setValor(resultset.getDouble("valor"));
                 produto.setStatus(resultset.getString("status"));
 
                 listagem.add(produto);
